+++ apps/message-broker-handler/src/config/message-broker.config.ts
@@ -3,11 +3,15 @@ import { AppEnvironmentVariables } from '../dtos';
 import { MessageBrokerConfiguration } from '@cali/message-broker-common';
 
 export const messageBrokerConfiguration = ({
-  MESSAGE_BROKER_URL,
-  MESSAGE_BROKER_USER,
-  MESSAGE_BROKER_PASSWORD,
+  MESSAGE_BROKER_DOMAIN: domain,
+  MESSAGE_BROKER_PORT: port,
+  MESSAGE_BROKER_PROTOCOL: protocol,
+  MESSAGE_BROKER_USER: user,
+  MESSAGE_BROKER_PASSWORD: password,
 }: AppEnvironmentVariables): MessageBrokerConfiguration => ({
-  url: MESSAGE_BROKER_URL,
-  user: MESSAGE_BROKER_USER,
-  password: MESSAGE_BROKER_PASSWORD,
+  domain,
+  port,
+  protocol,
+  user,
+  password,
 });