+++ apps/message-broker-handler/src/dtos/app-environment-variables.dto.ts
@@ -2,7 +2,13 @@ import { IsString } from 'class-validator';
 
 export class AppEnvironmentVariables {
   @IsString()
-  MESSAGE_BROKER_URL!: string;
+  MESSAGE_BROKER_DOMAIN!: string;
+
+  @IsString()
+  MESSAGE_BROKER_PORT!: string;
+
+  @IsString()
+  MESSAGE_BROKER_PROTOCOL!: string;
 
   @IsString()
   MESSAGE_BROKER_USER!: string;