+++ apps/video-analysis-mock/src/cli/cli.module.ts
@@ -11,24 +11,28 @@ import { MessageBrokerAlertModule } from '@cali/message-broker-alert';
 import { MessageBrokerPredictionLogModule } from '@cali/message-broker-prediction-log';
 import { MessageBrokerVideoGenerationModule } from '@cali/message-broker-video-generation';
 
+const {
+  MESSAGE_BROKER_DOMAIN: domain,
+  MESSAGE_BROKER_PORT: port,
+  MESSAGE_BROKER_PROTOCOL: protocol,
+  MESSAGE_BROKER_USER: user,
+  MESSAGE_BROKER_PASSWORD: password,
+} = process.env;
+
+const messageBrokerConfiguration = {
+  domain,
+  port,
+  protocol,
+  user,
+  password,
+};
+
 @Module({
   imports: [
     ConsoleModule,
-    MessageBrokerAlertModule.register({
-      url: 'amqp://localhost:5672',
-      user: 'guest',
-      password: 'guest',
-    }),
-    MessageBrokerPredictionLogModule.register({
-      url: 'amqp://localhost:5672',
-      user: 'guest',
-      password: 'guest',
-    }),
-    MessageBrokerVideoGenerationModule.register({
-      url: 'amqp://localhost:5672',
-      user: 'guest',
-      password: 'guest',
-    }),
+    MessageBrokerAlertModule.register(messageBrokerConfiguration),
+    MessageBrokerPredictionLogModule.register(messageBrokerConfiguration),
+    MessageBrokerVideoGenerationModule.register(messageBrokerConfiguration),
   ],
   providers: [AlertService, FramePerformanceService, VideoGenerationService],
 })