+++ libs/message-broker/common/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/dtos';
 export * from './lib/interfaces';
+export * from './lib/utils';