+++ libs/message-broker/common/src/lib/utils/generate-connection-url.ts
@@ -1,0 +1,11 @@
+import { MessageBrokerConfiguration } from '../dtos';
+
+export function generateConnectionUrl({
+  domain,
+  port,
+  protocol,
+  user,
+  password,
+}: MessageBrokerConfiguration) {
+  return `${protocol}://${user}:${password}@${domain}:${port}`;
+}