+++ libs/message-broker/common/src/lib/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './generate-connection-url';