+++ libs/nestjs/logger/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
+This file was generated using
+[@jscutlery/semver](https://github.com/jscutlery/semver).
 
 # [0.9.0](https://github.com/tractr/cali/compare/v0.8.0...v0.9.0) (2021-12-01)
 
@@ -16,7 +17,8 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 # [0.6.0](https://github.com/tractr/cali/compare/v0.5.0...v0.6.0) (2021-11-29)
 
-
 ### Features
 
-* add graphql push notification for alert feedback creation ([#59](https://github.com/tractr/cali/issues/59)) ([b01b40c](https://github.com/tractr/cali/commit/b01b40c017c939c4934293dc0ca67cb5ed01198b))
+- add graphql push notification for alert feedback creation
+  ([#59](https://github.com/tractr/cali/issues/59))
+  ([b01b40c](https://github.com/tractr/cali/commit/b01b40c017c939c4934293dc0ca67cb5ed01198b))